# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ModuleResponsePost": {
				"properties": {
					"moduleId": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"moduleId",
					"message"
				],
				"type": "object"
			},
			"Pick_Module.title-or-description-or-index_order-or-fk_course_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"fk_course": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"description",
					"index_order",
					"fk_course"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ModuleCreateRequest": {
				"$ref": "#/components/schemas/Pick_Module.title-or-description-or-index_order-or-fk_course_"
			},
			"DefaultSelection_Prisma._36_ModulePayload_": {
				"properties": {
					"fk_course": {
						"type": "number",
						"format": "double"
					},
					"id_module": {
						"type": "number",
						"format": "double"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"fk_course",
					"id_module",
					"index_order",
					"description",
					"title"
				],
				"type": "object"
			},
			"Module": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ModulePayload_",
				"description": "Model Module"
			},
			"ModuleResponsePut": {
				"properties": {
					"module": {
						"$ref": "#/components/schemas/Module"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"module",
					"message"
				],
				"type": "object"
			},
			"Pick_Course_class.class_id-or-title-or-index_order_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"class_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"index_order",
					"class_id"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ModuleClassSummary": {
				"$ref": "#/components/schemas/Pick_Course_class.class_id-or-title-or-index_order_"
			},
			"ModuleClassesResponse": {
				"properties": {
					"classes": {
						"items": {
							"$ref": "#/components/schemas/ModuleClassSummary"
						},
						"type": "array"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"module_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"classes",
					"index_order",
					"description",
					"title",
					"module_id"
				],
				"type": "object"
			},
			"ProgressResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"new_course_progress": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"new_course_progress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompleteClassRequest": {
				"properties": {
					"classId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"classId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ClassMaterial": {
				"properties": {
					"file_type": {
						"type": "string",
						"nullable": true
					},
					"file_url": {
						"type": "string",
						"nullable": true
					},
					"display_name": {
						"type": "string"
					},
					"file_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"file_type",
					"file_url",
					"display_name",
					"file_id"
				],
				"type": "object"
			},
			"CourseClassGetResponse": {
				"properties": {
					"materials": {
						"items": {
							"$ref": "#/components/schemas/ClassMaterial"
						},
						"type": "array"
					},
					"fk_module": {
						"type": "number",
						"format": "double"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"class_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"materials",
					"fk_module",
					"index_order",
					"description",
					"title",
					"class_id"
				],
				"type": "object"
			},
			"DefaultSelection_Prisma._36_Course_classPayload_": {
				"properties": {
					"fk_module": {
						"type": "number",
						"format": "double"
					},
					"class_id": {
						"type": "number",
						"format": "double"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"fk_module",
					"class_id",
					"index_order",
					"description",
					"title"
				],
				"type": "object"
			},
			"Course_class": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_Course_classPayload_",
				"description": "Model Course_class"
			},
			"CourseClassUpdateResponse": {
				"properties": {
					"class": {
						"$ref": "#/components/schemas/Course_class"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"class",
					"message"
				],
				"type": "object"
			},
			"Partial_CourseClassCreateRequest_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"fk_module": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CourseClassUpdateRequest": {
				"$ref": "#/components/schemas/Partial_CourseClassCreateRequest_"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ribbit_api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/ribbit/courses/{courseId}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"slug": {
											"type": "string"
										},
										"banner_url": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id_course": {
											"type": "number",
											"format": "double"
										},
										"modules": {
											"items": {
												"properties": {
													"classes": {
														"items": {
															"properties": {
																"is_completed": {
																	"type": "boolean"
																},
																"title": {
																	"type": "string"
																},
																"class_id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"is_completed",
																"title",
																"class_id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"index_order": {
														"type": "number",
														"format": "double"
													},
													"title": {
														"type": "string"
													},
													"module_id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"classes",
													"index_order",
													"title",
													"module_id"
												],
												"type": "object"
											},
											"type": "array"
										},
										"progress": {
											"type": "number",
											"format": "double"
										},
										"teacher_name": {
											"type": "string"
										}
									},
									"required": [
										"slug",
										"banner_url",
										"description",
										"title",
										"id_course",
										"modules",
										"progress",
										"teacher_name"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Curso não encontrado"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCourse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"course": {
											"properties": {
												"slug": {
													"type": "string"
												},
												"banner_url": {
													"type": "string"
												},
												"description": {
													"type": "string"
												},
												"title": {
													"type": "string"
												},
												"id_course": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"slug",
												"banner_url",
												"description",
												"title",
												"id_course"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"course",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Curso não encontrado"
					},
					"409": {
						"description": "Slug já em uso"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"slug": {
										"type": "string"
									},
									"banner": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteById",
				"responses": {
					"204": {
						"description": "Deletado com sucesso"
					},
					"404": {
						"description": "Curso não encontrado"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/ribbit/modules": {
			"post": {
				"operationId": "CreateModule",
				"responses": {
					"201": {
						"description": "Criado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ModuleResponsePost"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao criar módulo"
					}
				},
				"tags": [
					"Modules"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ModuleCreateRequest"
							}
						}
					}
				}
			}
		},
		"/ribbit/modules/{module_id}": {
			"delete": {
				"operationId": "DeleteModule",
				"responses": {
					"204": {
						"description": "Deletado"
					},
					"404": {
						"description": "Módulo não encontrado"
					}
				},
				"tags": [
					"Modules"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "module_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateModule",
				"responses": {
					"200": {
						"description": "Atualizado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ModuleResponsePut"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao atualizar módulo"
					},
					"404": {
						"description": "Módulo não encontrado"
					}
				},
				"tags": [
					"Modules"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "module_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ModuleCreateRequest"
							}
						}
					}
				}
			}
		},
		"/ribbit/modules/{module_id}/classes": {
			"get": {
				"operationId": "GetModuleClasses",
				"responses": {
					"200": {
						"description": "Classes encontradas",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ModuleClassesResponse"
								}
							}
						}
					},
					"404": {
						"description": "Módulo não encontrado"
					}
				},
				"tags": [
					"Modules"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"aluno",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "module_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ribbit/progress": {
			"post": {
				"operationId": "CompleteClass",
				"responses": {
					"201": {
						"description": "Criado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProgressResponse"
								}
							}
						}
					},
					"400": {
						"description": "Erro de validação"
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"Progress"
				],
				"security": [
					{
						"bearerAuth": [
							"aluno",
							"prof",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompleteClassRequest"
							}
						}
					}
				}
			}
		},
		"/ribbit/progress/{class_id}": {
			"delete": {
				"operationId": "RemoveClassCompletion",
				"responses": {
					"200": {
						"description": "Conclusão removida com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProgressResponse"
								}
							}
						}
					},
					"404": {
						"description": "Registro não encontrado"
					}
				},
				"tags": [
					"Progress"
				],
				"security": [
					{
						"bearerAuth": [
							"aluno",
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "class_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/ribbit/classes/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseClassGetResponse"
								}
							}
						}
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"CourseClass"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCourseClass",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseClassUpdateResponse"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao atualizar a aula"
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"CourseClass"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CourseClassUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCourseClass",
				"responses": {
					"204": {
						"description": "Aula deletada com sucesso"
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"CourseClass"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}